Updated openstack/openstack Project: openstack/cinder  091574f57cebdaa3e7d3120d18441ef7713d7c5a
Move oslo.serialization to oslo_serialization namespace

This is the fifth in a series of changes to move to using
the new oslo_<library> namespace that is being used for
oslo libraries.

There is currently a shim in place that is allowing the old
oslo.<library> imports to work, but we need to be prepared for
when the shims go away.  Thus, we need patches like this one to
move to the new namespace.

This patch also updates our hacking check to ensure that no instances
of oslo.utils sneak back in.

Change-Id: Ida0f9512a816d45ec5cae4dc4da4dbc923a67cf4
Partial-bug: 1409733