Updated openstack/openstack Project: openstack/python-novaclient  0ae527665a7db5087216e8cb40da7d21d51a7848
Use TCP Keep-Alive on the socket level

There is not a way to pass the socket options to the HTTPAdapter upon
creation so we have to sub-class it and override the init_poolmanager
method.  This also requires at least python-requests 2.4.0 but that
has 2 severe bugs that were fixed in 2.4.1. If we try to fix this
without a hard lower limit, we will not be able to properly set these
options on the socket at creation time.

Change-Id: I06e0d2c67d3197607e5f23f623c8fca69e1b23d7
Closes-bug: 1323862