Updated openstack/openstack Project: openstack/taskflow  e417914d49458a4f0fd231497f6a4c1200cfecb2
Make all/most usage of type errors follow a similar pattern

To make it easy to add new type errors and to make the existing
ones have a common string pattern adjust the current type errors
to contain at least the following string format: '%s' (%s) where
these two places will be filled in with the object of the wrong
type and the type of that object. This information is useful when
analyzing the exception (by the user) to know exactly what they
passed in and what type it was.

This convention is not maintained where it would interpolate large
text blobs (such as in binary encoding/decoding and json decoding).

Change-Id: Id84b0e7ce684a543cc407b15016e77804e6f03ed