Updated openstack/openstack Project: openstack/nova  159e68eb1ab28e271ca35b8344d905f1099f7d61
xenapi: fix structure of data reported for cpu_info

The 'cpu_info' field in the dict from get_available_resource
is supposed to provide information on the physical CPU model
used in the host, including arch, model, vendor, feature flags
and topology. The XenAPI Driver was simply returning the
raw 'host_cpu_info' dict it got back from XenSever. This
includes nearly all the desired information, but represented
in a different structure to what the libvirt driver has
provided.

This commit updates it to convert the data into a format
that is somewhat consistent with the libvirt driver.

Cleanup before Blueprint: resource-objects
Change-Id: Ic21990e717bc2ef623b5feeb3eee3bd758bbe8af