Updated openstack/openstack Project: openstack/nova  733ef1176def03a871765a51cd16c661d3175935
hyperv: use standard architecture constants for CPU model

When reporting the 'cpu_info' field in the get_available_resource
method, one of the fields is the architecture. The hyperv driver
has mistakenly used printable architecture names, as opposed to
the canonical architecture names. Switch to use the compute.arch
constants.

Although this changes the data reported by hyperv, it should not
have a functional impact since only the libvirt driver is relying
on the precise format of this data. For hyperv it is merely
informational.

Cleanup before Blueprint: resource-objects
Change-Id: I6bfbfc829f45dee49e1f2eeaa0490068da26ccb2