Updated openstack/openstack Project: openstack/cinder  a8499b24a99f722af94e05ea63f5b75fcf9ba71b
Add a generic retry decorator to cinder/utils

Retries are something that we use in a number of
places, and could probably use in a few more.  Rather
than continue writing custom retry code or silly loops
let's add a generic retry decorator class that can be
initialized and used as needed.

This patch leverages the retrying library that's already
in OpenStack Requirements and used in other places.  We
just add a wrapper around it for some logging and add
a bit of error handling incase somebody sets retries to 0

Co-Authored-By: Josh Harlow <harlowja@yahoo-inc.com>

Change-Id: I629373d3c75be591ccde5608c8c85f2debd47447