Updated openstack/openstack Project: openstack/nova  4511c8e6572900a05cc745ac8ffa59a12ee134fd
libvirt: Consider CPU pinning when booting

Consider CPU pinning request when booting an instance and make sure
that the proper XML is generated. If there was some pinning information
generated as part of the proposed NUMA configuration, this will make
it more specific as it will pin to the exact pCPU core as opposed to
the set that belongs to a host's NUMA node.

Blueprint: virt-driver-cpu-pinning

Change-Id: Iec0107034366eef1df9ce49dc90b14e3ab2a270a