Updated openstack/openstack Project: openstack/ceilometer  6427efc6416dadf9d1ac96490d00bb2452cffc23
[DB2 nosql] Create TIMESTAMP type index for 'timestamp' field

Currently, when doing ceilometer-dbsync, ceilometer will create an
index for 'timestamp' field of meter collection. But the datatype
of the field is set to 'VARCHAR' which is not correct.

The patch insert a document with 'timestamp' before creating the
index of 'timestamp', then DB2 nosql will create the correct index
type for 'timestamp' field.

Change-Id: I69dd2fc8d552f5dbb2485c2a5eab55b3bee71e29
Closes-Bug: #1411492