Updated openstack/openstack Project: openstack/oslo.messaging  eb9251173c30013616ae109b5b945e87e5347e18
Expose _impl_test for designate

The previous attempt to do this in
Ide40b293c9b9c76aae094950720cead4179ea883 created a new module
oslo/messaging/notify/_impl_test.py which imported the symbols from
oslo_messaging/notify/_impl_test.py. That doesn't actually work, because
the functions to reset the notifications modify the copy of the symbol
in the oslo_messaging package, but the designate tests are still looking
at the copy in oslo/messaging.

This change imports a copy of the module from oslo_messaging in the
oslo/messaging namespace to make certain that the same module object,
and therefore the same global data structure, is used no matter which
path the code is accessed.

The designate team still needs to fix their fixture, but this will let
us release oslo.messaging without breaking their tests in the mean time.

Partial-Bug: #1412812

Change-Id: I629ac57ae1e74bb07e5e98ee2fbc21f8421c3c24