Updated openstack/openstack Project: openstack/ceilometer  889cc514832563c45c6798120448a4b67a6991f5
Use alarm's evaluation periods in sufficient test

Currently, we use constant value quorum=1 to check if there are enough
datapoints, however, this is not quite right for an alarm rule.

Image evaluation periods is set to, for i.e., 3 for an instance on
cpu_util greater or equal than 80%. Here are the cases which current
may not work as expected:

1. when system start or instance is just created, we may only get one
or two samples for the instance
2. when system is somewhere broken, or an instance is restarted (after
being shutoff), sample may fail to be collected in some time, so we only
get one or two sample in that time range

We want to avoid a spurious data peak, for example, instance cpu_util can
be 50%, 50%, 50%, 90%, in such case, alarm will not be triggered, but if
instance cpu_util is None, None, None, 90%, current code will think alarm
should be triggered, which is not consistent and may confuse end users.

This patch will put alarm to insufficient data when datapoints are less
than evaluation periods.

Change-Id: Ie64a537434493a5965c8e9e165cf028d57689da2
Closes-Bug: #1380216