Updated openstack/openstack Project: openstack/tempest  915182016acb97dd50e6a7cc1b198464a1853f6d
Allow to specify the list of sahara enabled plugins

The list of plugins can be tuned by cloud administrator, so make
sure that the list of checked plugins is not hardcoded.
The default list of plugins matches the one currently enabled.

This commit introduces also a new section for data_processing enabled
features in the tempest configuration.

Change-Id: Ie29aeef2df605cfe840e053bc57ff980e9332636