Updated openstack/openstack Project: openstack/heat  51a831f2cb2404207a279f2ccbcf7d5059f10e7b
proper handling of unsupported request methods

This change adds support for responding to unsupported methods for all
routes. Invoking a route with an unsupported request method returns a 405
status code, and the response includes the 'Allow' header listing the
request methods supported by the requested URL. The OPTIONS request method
is also automatically supported for all URLs, with the response
returning the 'Allow' header. The way routes are registered has been
changed so that unsupported methods can be detected automatically.

Change-Id: I2de28e0fc6cd35ed060395405aa3770b1dc73376
Closes-Bug: 1367057
APIImpact
DocImpact