Updated openstack/openstack Project: openstack/nova  5bdb96abce5c74d1cbf904e7f3d7b4a8b9449949
Add API schema for v2.1 createImage API

By defining the API schema, it is possible to separate the validation
code from the API method. The API method can be more simple.
In addition, a response of API validation error can be consistent for
the whole Nova API.

Partially implements blueprint v2-on-v3-api

Change-Id: I6b1405616016f3a8c751649cb7610ff37f78424a