Updated openstack/openstack Project: openstack/sahara  a8f336377736c11844846e8da579365fdd77db6b
Follow the argument order specified in spark-submit help

The command issued by Sahara to run jobs with spark-submit does
not put the application jar in the right place according to the
help text of spark-submit. This does not make jobs fail, but it
is good to be consistent in case something changes.

Change-Id: I50c2a969e4f747820c06d5dba39b6a8442bb5c30
Closes-Bug: #1410247