Updated openstack/openstack Project: openstack/trove  c7cb6ef711ae0da603ae145cb5ca47482701178a
Support string flavor IDs

This allows operating on Nova flavors with non-integer ids
by adding a new str_id field to the view that always contains the
unicode flavor id.  The current id field will now contain a value
only when the flavor's id can be cast to int, otherwise it will be
None.  Validation of flavor id when querying a flavor or creating an
instance has been updated to include non-empty strings in addition to
integers.

This will require a patch to python-troveclient to properly fallback
to str_id in absence of the integer id:

    https://review.openstack.org/#/c/123301/

Change-Id: Ie9cfefc6127bc76783cdf9668636342d7590b308
Closes-bug: #1333852