Updated openstack/openstack Project: openstack/nova  e8e37a52896aea9b557ad152bad24f52019965c8
Check for LUKS device via 'isLuks' subcommand

Determination of whether or not a device uses LUKS for encryption
is currently performed by parsing the output of the 'luksOpen'
command. This approach is brittle and does not handle non-English
locales.

This change replaces the existing approach with the 'isLuks' command
to determine if the device uses LUKS. In addition, test cases are
added to provide better coverage of the attach_volume operation
for encrypted volumes.

Closes-Bug: 1372108

Change-Id: I99f905f8f5c7d46ecc826fa0da5d36f54f1107dc