Updated openstack/openstack Project: openstack/python-keystoneclient  0eab8ff9204f946d1d51caa5ed076dc954e67e2b
Update requests-mock syntax

With requests-mock 0.5 we can do away with register_uri(method, ..) and
just use the method name as a function. This feels much cleaner and
looks more like requests syntax.

In the same update there is query string decoding included in the
history, so we no longer have to manually parse the query strings out of
the url.

Change-Id: I43d31576d15b4be72350bebf00733c08a7fb3e6c