Updated openstack/openstack Project: openstack/heat  4947b0a233a5a1413c030f8fffa311ff01438852
Use keystoneclient auth plugins

This is the first step in converting to using keystoneclient auth
plugins. It moves responsibility for controlling authentication away
from the keystoneclient and onto the equivalent functions provided by
the context. The existing functions are left for now with redirects to
the context.

Have the context create an authentication plugin for the user
credentials in the same way that keystoneclient used to. This involves
updating a number of tests as we need to stub authentication information
from the context and not the client.

Change-Id: Ifeb5c99387aec0eb5dce3f2744388c948d78a2bb