Updated openstack/openstack Project: openstack/python-heatclient  4b5b84500a5c5932de341ff62a6d2c378b6eb91b
output-show format by json or raw

By default the output-show command formats the output value as
serialized json, which is inconvenient if the value is a simple
multi-line string.
This change adds a -F, --format option to specify json or raw for
the format of the output value. The default retains the current
json behaviour. List and map values use json formatting regardless
of the selected format.
Change-Id: I07c8b2c08daae4ac9f56b06863461ee22e6ad0d8