Updated openstack/openstack Project: openstack/keystone  32d2267b2cfb004965505267b4e060a49e15c643
Move test_utils to keystone/tests/unit/

This commit moves the unit tests that are in the test_utils.py module to
keystone/tests/unit/ directory. There was also a duplicate module that
tested keystone.common.utils so those tests were refactored so that they all
live in the same test module.

Change-Id: Ie1d84ae50faf25319099a224639ed27826b612ad