Updated openstack/openstack Project: openstack/python-keystoneclient  e9dc0b953ab00fbc3413537355fc00048b98a1f0
Surface the user_id and project_id beyond the plugin

Having the user_id and project_id exposed in the plugin is a good first
step however we don't really expect the user to be interacting with the
plugins directly often - particularly as you need to pass session to the
methods.

Exposing get_user_id and get_project_id on the session and the adapter
in this way is very similar to the way we expose get_token and
get_endpoint on the session and adapter for use higher up.

Related-Bug: #1364724
Change-Id: If2f868c3ddc19133f18446e74f8e1b560a4798fa