Updated openstack/openstack Project: openstack/neutron  d0da02070c79127ff5c50372892b5009cca473df
Fixes blocking of VRF config in Arista L3 Plugin

This fix removes the restriction that Arista HW,
when configured in MLAG (redunancy) mode, is not
allowed to configure in VRF mode.

Additionally, a minor fix to log message is made
as the correct message is logged when an
exception is hit.

Change-Id: Icc1e9cc81d20c6aa19ed0429bdbc781e4a11f224
Closes-bug: 1411016