Updated openstack/openstack Project: openstack-dev/devstack  ea8fb0c246339ae04d1791e7c3743264cecb9986
Use LVMVolumeDriver instead of LVMISCSIDriver

The Cinder LVMISCSIDriver is now deprecated. As a result, default
settings in devstack are giving warnings in the Cinder volume serivce.
The LVMVolumeDriver now handles all cases, by looking at the
iscsi_helper. This will use that driver instead, which will stop the
deprecation warnings.

Closes-bug: #1413761
Change-Id: Ifbb9ce45694095ff9e30f3ca4c3859a07de8df73