Updated openstack/openstack Project: openstack/nova  422b22eaf79c04e42e1b2257b7ab1feff54152bc
Network: correct VMware DVS port group name lookup

Commit 709127cabed602d6868dedf0bef7ca9547a86e06 added support
for DVS network backends. The original design included the
name of the network in the port group. The name was excluded due
to VC limitations (maximum of 80 characters for a port group name).
This prevents instances from being booted as there is no match
for the port group.

The fix ensures that we only use the Neutron network UUID.

Change-Id: I7e3223f88c4f477116ac493eb45466ea1c0a29a9
Closes-bug: #1415461