Updated openstack/openstack Project: openstack/neutron  274c66f291f40224e3f16dcdf02b17ccb1196924
Make the interface driver available to the router classes

Ultimately, it will only be the routers that need access to the
interface driver and the agent won't need to use it for anything.
However, it still makes sense for the agent to initialize it once and
pass it to each of the routers as they're created.

Multiple patches with multiple authors will be created to depend on
this addition.

Change-Id: I0ec75d731d816955c1915e283a137abcb51ac232
Partially-Implements: bp/restructure-l3-agent