Updated openstack/openstack Project: openstack/neutron  06233db2e790891c20c2b9a72a2b38146d928b60
Allow 'max_l3_agents_per_router' to be set to '0'

Currently the field 'max_l3_agents_per_router' from
'neutron.conf' cannot be set to '0' even though the comments
and code indicate it is be supported. The value
means 'unlimited' agents per router is allowed on HA routers.

This issue is a regression that was caused by:
7da314434e445ce3a6f3642c784954ef61154b7f

This patch adds a special handling for this value when validating
the config. When a value of '0' is used, the further validation
of max value is skipped.

Change-Id: Iac85768b350ee16c34893218738974a2692202aa
Closes-Bug: #1414548