Updated openstack/openstack Project: openstack/os-net-config  c86a76e13ea07005e31890695c220081b0ed816d
Remove deprecated references to oslo-incubator

Purge references to modules which have been deprecated by various oslo
libraries.

Requires changes to use oslo.utils and oslo.concurrency, which are
added to the requirements, syncing with the latest in global requirements
so we can also move to the un-namespaced oslo package names.

Change-Id: Idacb71b7871330e3b3fabf4a926a5b8987614c4d