Updated openstack/openstack Project: openstack/heat  d2bd899268c94d4a43395f912aee511f81e49c10
Store auth_url and region_name when trusts are enabled

We don't currently store auth_url, which causes problems when creating
autoscaling groups containing servers configured for software config,
because we use context.auth_url to populate the auth_url in the server
metadata, which is needed by the deployments for signalling.

Also adding region_name, as this seems likely to be needed for correct
operation in a multi-region environment.

Closes-Bug: #1415901
Change-Id: I550d03ea1b6985ba0777bf5ca52c79a8e8e60f5a