Updated openstack/openstack Project: openstack/horizon  a44886b1d0538bb128e759738388264ea041ae8b
password visibility icon not shown for first page

In the initial page (http://<ip>:<port>) the
password field not showing the password visibility
icon.

If we go to login page (http://<ip>:<port>/auth/login)
the password visibility icon is shown properly.

Now both the address are using the same template so
the password visibility icon shown in the initial
page also and the unused template is deleted.

Change-Id: Ia8063c54143c5394c927cf7aea7c15c3e9373dd0
Closes-Bug: #1408220