Updated openstack/openstack Project: openstack-dev/devstack  b9a7d3b5e4f01d932709633f2b5b9461ecdbd0cb
Revert "Exit after printing error about RPC config"

This reverts commit c1dbf10562796c7006fd0a66c6cb39a1b990c8c1.

Reverting because it is valid to run devstack without enabling an rpc
backend. For example when you run devstack on a multinode deployment
compute hosts probably shouldn't run rabbit and instead will just talk
to the rabbit server set up on the controller node. As a result we
should not die when no rpc backends are set.

Change-Id: If9a31e031552f4161d42094fa960221b4d9e893c