Updated openstack/openstack Project: openstack/nova  4cdb032dd1e362df73b78560f34c1d31ab59c40a
don't assert_called_once_with with a real time

This test failed in the gate because there was a 10ms delay between
the time allocation and the assert_called_once_with (so the time
parameter did not match). Instead use mock.ANY for the time field, the
payload is not important.

Change-Id: Ibf7041280c7c6cc50334bc6b8489c4c1c38ecb24