Updated openstack/openstack Project: openstack/heat  1fc5d2b67e737805292052c491c85da01c842ab5
Don't do snapshot when other action in progress

According to the referenced bug report, snapshot and restore operations
are not checking the stack status. The result is that some duplicate or
orphan VMs left. This patch adds a check in snapshot operation that will
reject a snapshot request when it seems unsafe.

partial-bug: 1412952
Change-Id: I5b5da312ac08fd5f15418dd6cf9670463180d7ad