Updated openstack/openstack Project: openstack/neutron  6a797f354eb4ba936b80603f7cc01a2fe80446fd
Implements the ProcessMonitor in the l3_agent

The ProcessMonitor class will watch over spawned external processes,
taking the administrator configured action in the case of any
of the external processes dying unexpectedly.

It covers both the neutron-ns-metadata-proxy for non-ha routers
and the IPv6 radvd external processes. Keepalived +
neutron-ns-metadata-proxy needs to be covered in a second follow up
patch when neutron-ns-metadata-proxy is handled by the l3-agent
(instead keepalived) in the ha-routers.

Implements: blueprint agent-child-processes-status

Change-Id: Id6cc4786d837b96c61429d51485bc86ae37872cb