Updated openstack/openstack Project: openstack/keystone  08afed58d3f2c209875a33bcbf56e4ad5caedd0e
Make unit tests call the new resource manager

This is the part of the more comprehensive split of
assignments, which rationalizes both the backend and controllers.
In order to make this change easier for reviewers, it is divided
into a number of smaller patches.

Previous patches have:

- Moved role management into its own manager and drivers
- Fixed incorrect doc strings for grant driver methods
- Updated controllers to call the new role manager
- Updated unit tests to call the new role manager
- Refactored the assignment manager and drivers enabling
  projects/domains to be split out
- Fixed incorrect comment about circular dependency between
  assignment and identity
- Moved the logically separated project and domain
  functionality into their own manager/backend (called resource).
- Removes unused pointer to assignment from identity driver
- Uddated controllers and managers to call the new resource
  manager

This patch updates all the unit tests to call the new resource
manager to access projects and domains.

Future patches will:
- Split the assignment controller, giving projects/domains
  their own controller

Partially implements: bp pluggable-assignments
Change-Id: I503fbcfa9fb141a3769aa831b8c6a429ba37c498