Updated openstack/openstack Project: openstack/neutron  cff0e6bce31fa73d8a66556cf6ffc1d3fcb05152
Thining Arista ML2 driver from neutron tree

Moving all the back-end drivers for the Arista
ML2 driver. The DB Models and Config is
left in the tree. Rest all is moving to
networking-arista project.

Reorganize the code to ensure a smoother transition
to a decomposed Arista ML2 driver.

The code is being moved in [1].

[1] https://review.openstack.org/#/c/148682/

Partial-bug: 1412577
Related-blueprint: core-vendor-decomposition

Change-Id: Ibbd82d7a18fdb8d6b8d4bfa7bcafe8a2d91a20e1
Co-Authored-By: Armando Migliaccio <armamig@gmail.com>