Updated openstack/openstack Project: openstack/horizon  e9615f03b302e9a91d0cbbc10f46cd566e67fc5e
Enables the editing of Allocation pools in edit subnet

This patch enables the editing of the allocation pools in the edit
subnet form. It removes the allocation pools hidden field because it
is defined in the base class. So this make allocation pools field
visible in the edit form with pre-populated values.

Change-Id: I8d0c35f0694cc5a651747209bee6d38c1d9e26a9
Closes-bug: #1410173