Updated openstack/openstack Project: openstack/horizon  ba15ee663b213c59ce0efd6ae2c0f1a01fd30e7c
added flavor id in instance detail page

in instance table, we can filter out the
instances by using the flavor id.

but there is no way to get the flavor id
in horizon for non admin users.

so adding the flavor id in instance detail
and popover flavor detail table.

Change-Id: I4995d9181a47bb7e15ce8b2c7b4252e389d91c71
Closes-Bug: #1415825