Updated openstack/openstack Project: openstack-infra/storyboard  7e62b72503737f9429dac52c6cf42316f0e4b44d
Email Outbox

An email outbox, stored in the plugin's working directory, which
will be used to store our outgoing messages before they are handed
off to an SMTP server. This is done because we cannot anticipate the
size of the outgoing message (as it could contain a large number of
digest events), and do not want to slow down the email build processor.

While we use mailbox.Maildir as a base class, I overrode
the name generation to prevent conflicts warned against in the
documentation (see link). I also overrode the factory to circumvent
an incompatible factory in use in python2.7.

A context-manager for the outbox is also provided, to make sure that
we always clean up after ourselves.

https://docs.python.org/2/library/mailbox.html#maildir

Change-Id: I94c8b22a25a3054d507a04b84ce1043975fc13c2